--- a/lib/index.js
+++ b/lib/index.js
@@ -21,25 +21,27 @@ const validateSource = (source) => {
     if (!spec) {
       return next(null);
     }
-    Joi.validate(req[source], spec, options, (err, value) => {
-      if (value !== undefined) {
-        const descriptor = Object.getOwnPropertyDescriptor(req, source);
-        /* istanbul ignore next */
-        if (descriptor && descriptor.writable) {
-          req[source] = value;
-        } else {
-          Object.defineProperty(req, source, {
-            get () { return value; }
-          });
-        }
+    const result = Joi.validate(req[source], spec, options);
+    const value = result.value;
+    const err = result.error;
+
+    if (value !== undefined) {
+      const descriptor = Object.getOwnPropertyDescriptor(req, source);
+      /* istanbul ignore next */
+      if (descriptor && descriptor.writable) {
+        req[source] = value;
+      } else {
+        Object.defineProperty(req, source, {
+          get () { return value; }
+        });
       }
-      if (err) {
-        err[CELEBRATED] = true;
-        err._meta = { source };
-        return next(err);
-      }
-      return next(null);
-    });
+    }
+    if (err) {
+      err[CELEBRATED] = true;
+      err._meta = { source };
+      return next(err);
+    }
+    return next(null);
   };
 };
 
@@ -47,6 +49,22 @@ const validateHeaders = validateSource('headers');
 const validateParams = validateSource('params');
 const validateQuery = validateSource('query');
 const validateBody = validateSource('body');
+const maybeValidateBody = (config, callback) => {
+  const method = config.req.method.toLowerCase();
+
+  if (method === 'get' || method === 'head') {
+    return callback(null);
+  }
+
+  validateBody(config, callback);
+};
+
+const REQ_VALIDATIONS = [
+  validateHeaders,
+  validateParams,
+  validateQuery,
+  maybeValidateBody
+];
 
 const isCelebrate = (err) => {
   if (err != null && typeof err === 'object') { // eslint-disable-line eqeqeq
@@ -67,18 +85,7 @@ const celebrate = (schema, options) => {
     rules.set(key, Joi.compile(schema[key]));
   }
   const middleware = (req, res, next) => {
-    Series(null, [
-      validateHeaders,
-      validateParams,
-      validateQuery,
-      (config, callback) => {
-        const method = config.req.method.toLowerCase();
-        if (method === 'get' || method === 'head') {
-          return callback(null);
-        }
-        validateBody(config, callback);
-      }
-    ], {
+    Series(null, REQ_VALIDATIONS, {
       req,
       options: joiOpts,
       rules